refactor: MVC 패턴에 맞게 리팩토링
